+++ gluecodium/src/test/resources/smoke/visibility_private/output/android/com/example/smoke/PrivateFields.java
@@ -1,0 +1,19 @@
+/*
+ *
+ */
+package com.example.smoke;
+import android.support.annotation.NonNull;
+public final class PrivateFields {
+    @NonNull
+    public String publicField;
+    @NonNull
+    private String privateField;
+    public PrivateFields(@NonNull final String publicField) {
+        this.publicField = publicField;
+        this.privateField = "nonsense";
+    }
+    public PrivateFields(@NonNull final String publicField, @NonNull final String privateField) {
+        this.publicField = publicField;
+        this.privateField = privateField;
+    }
+}